# perf(treesitter): add `on_range` callback to decoration provider