--- a/src/nvim/decoration_defs.h
+++ b/src/nvim/decoration_defs.h
@@ -149,6 +149,7 @@ typedef struct {
   LuaRef redraw_buf;
   LuaRef redraw_win;
   LuaRef redraw_line;
+  LuaRef redraw_range;
   LuaRef redraw_end;
   LuaRef hl_def;
   LuaRef spell_nav;