… runtime/doc/api.txt
… runtime/doc/news.txt
… runtime/doc/treesitter.txt
… src/nvim/drawline.c
… test/benchmark/decor_spec.lua
… test/benchmark/treesitter_spec.lua
… test/functional/treesitter/query_spec.lua
… test/functional/ui/decorations_spec.lua