Added `on_range` callback to decoration provider that takes starting and ending rows and columns. Updated treesitter highlighter to use `on_range` instead of `on_line`. This improves performance for long lines.

The callback might be called multiple times for the same line with different offsets. This is done because we can't predict how many characters will be visible (because of `conceal`, which can be added by the callback itself if I understand correctly, as well as multibyte and combining characters). 

For now, the callback is called every 500 bytes and at the end of the line.

All bounds can be outside the buffer and in the middle of a character.

### What needs to be done:

- [ ] Some tests are failing. One of the reasons is because highlights are sometimes added when they are outside the current range ([see comment](https://github.com/neovim/neovim/pull/30869#issuecomment-2430570060)). 

For example, `supports language injections` test fails because the injection tree iterator is created at the end of the first line, and adds its first match (on the second line) before highlights from the main tree.  Before, the iterator for the injection would be created on the second line, after highlights from the main tree for that line were added. There are probably more issues similar to this.

https://github.com/neovim/neovim/blob/2833925cfc688786759d6a980a1ad62b62d20570/test/functional/treesitter/highlight_spec.lua#L860-L863

- [ ] The API needs to be updated properly, without `tslua_push_querycursor2()` and `iter_captures2()`.

- [ ] Find a better heuristic for the callback range.