@luukvbaal 2024-11-30
Cool! Idk if you had seen the conceal_lines PR at all but it currently contains a commit that will affect this (or the other way around), and seemingly force the `on_range` callback here to still iterate over a table of all marks in a line :/ [`b125b57` (#31324)](https://github.com/neovim/neovim/pull/31324/commits/b125b5771999889d36aca6a66b5f94e21832e295). Although perhaps `on_range` could remove handled entries from the prepared highlight state to avoid that.