@vanaigr 2024-12-01
Fixed the failing functional test.

This doesn't mean the bug described in the [comment](https://github.com/neovim/neovim/pull/30869#issuecomment-2430570060) is fixed.