@vanaigr 2024-12-01
Now initial range is calculated from the number of cells remaining to be drawn.

Is using `w_width_inner` correct, or should it be `w_width`? Are they different in multigrid?