@vanaigr 2024-12-01
For `nowrap`, if EOL is not visible, the transition range (from one line to the other) is not called. This means the iterator will be recreated.