@zeertzjq 2024-12-01
`w_width_inner` should be used for the rendering of text in a window