@vanaigr 2024-12-03
I should also mention that there's another possible issue:

Injection highlights might get obscured by main highlights if both overlap have the same priority. This can happen if main highlight is positioned after injection highlight, and they are added in separate ranges. Because insertion order takes priority over position, "older" injection highlight would be hidden by "newer" main highlight. If they were added in the same range, the order would be reversed.

This is not a new issue, since the same applies to lines. If injection highlight spans multiple lines, and main highlight gets added e.g. on the second line.

I wasn't able to come up with any examples of this behavior, though.