@vanaigr 2024-12-07
Added benchmarks.