@vanaigr 2024-12-07
here are the results:

### Before

```
RUN      T2 treesitter perf can redraw the beginning of a long line with wrapping:
Total min, 25%, median, 75%, max:
        17.6ms, 21.3ms, 21.9ms, 22.8ms, 25.7ms
2370.00 ms OK
RUN      T3 treesitter perf can redraw the middle of a long line with wrapping:
Total min, 25%, median, 75%, max:
        17.5ms, 19.6ms, 20.2ms, 21.2ms, 25.7ms
2194.00 ms OK
RUN      T4 treesitter perf can redraw the end of a long line with wrapping:
Total min, 25%, median, 75%, max:
        14.1ms, 17.7ms, 18.2ms, 19.1ms, 24.8ms
1995.00 ms OK
RUN      T5 treesitter perf can redraw the beginning of a long line without wrapping:
Total min, 25%, median, 75%, max:
        14.5ms, 18.1ms, 18.6ms, 19.5ms, 24.5ms
2009.00 ms OK
RUN      T6 treesitter perf can redraw the middle of a long line without wrapping:
Total min, 25%, median, 75%, max:
        13.9ms, 18.5ms, 19.1ms, 20.0ms, 21.7ms
2026.00 ms OK
RUN      T7 treesitter perf can redraw the end of a long line without wrapping:
Total min, 25%, median, 75%, max:
        15.4ms, 17.7ms, 18.2ms, 19.0ms, 21.5ms
1940.00 ms OK
```

### After

```
RUN      T2 treesitter perf can redraw the beginning of a long line with wrapping:
Total min, 25%, median, 75%, max:
        0.9ms,  0.9ms,  1.0ms,  1.4ms,  4.2ms
293.00 ms OK
RUN      T3 treesitter perf can redraw the middle of a long line with wrapping:
Total min, 25%, median, 75%, max:
        1.3ms,  1.4ms,  1.5ms,  1.8ms,  5.7ms
324.00 ms OK
RUN      T4 treesitter perf can redraw the end of a long line with wrapping:
Total min, 25%, median, 75%, max:
        1.6ms,  1.7ms,  1.7ms,  2.1ms,  4.7ms
344.00 ms OK
RUN      T5 treesitter perf can redraw the beginning of a long line without wrapping:
Total min, 25%, median, 75%, max:
        0.1ms,  0.1ms,  0.1ms,  0.2ms,  2.8ms
180.00 ms OK
RUN      T6 treesitter perf can redraw the middle of a long line without wrapping:
Total min, 25%, median, 75%, max:
        0.6ms,  0.6ms,  0.6ms,  0.6ms,  3.3ms
216.00 ms OK
RUN      T7 treesitter perf can redraw the end of a long line without wrapping:
Total min, 25%, median, 75%, max:
        0.9ms,  0.9ms,  1.0ms,  1.0ms,  4.1ms
256.00 ms OK
```

`Before` becomes faster at the end of the line because there are less decorations to iterate (since the ones before the cursor are removed).