@vanaigr 2024-12-07
Here are the results with #30869:

### Before

```
RUN      T2 treesitter perf can redraw the beginning of a long line with wrapping:
Total min, 25%, median, 75%, max:
        13.5ms, 17.6ms, 18.3ms, 19.5ms, 27.0ms
1976.00 ms OK
RUN      T3 treesitter perf can redraw the middle of a long line with wrapping:
Total min, 25%, median, 75%, max:
        15.2ms, 18.0ms, 18.4ms, 19.4ms, 27.8ms
1982.00 ms OK
RUN      T4 treesitter perf can redraw the end of a long line with wrapping:
Total min, 25%, median, 75%, max:
        13.5ms, 18.6ms, 19.6ms, 20.6ms, 23.5ms
2046.00 ms OK
RUN      T5 treesitter perf can redraw the beginning of a long line without wrapping:
Total min, 25%, median, 75%, max:
        15.7ms, 18.2ms, 18.6ms, 19.7ms, 22.0ms
1989.00 ms OK
RUN      T6 treesitter perf can redraw the middle of a long line without wrapping:
Total min, 25%, median, 75%, max:
        13.9ms, 18.3ms, 18.8ms, 19.7ms, 22.3ms
1985.00 ms OK
RUN      T7 treesitter perf can redraw the end of a long line without wrapping:
Total min, 25%, median, 75%, max:
        13.5ms, 18.1ms, 18.7ms, 19.7ms, 22.4ms
1977.00 ms OK
```

### After

```
RUN      T2 treesitter perf can redraw the beginning of a long line with wrapping:
Total min, 25%, median, 75%, max:
        0.8ms,  0.8ms,  1.0ms,  1.3ms,  4.0ms
300.00 ms OK
RUN      T3 treesitter perf can redraw the middle of a long line with wrapping:
Total min, 25%, median, 75%, max:
        1.1ms,  1.2ms,  1.4ms,  1.6ms,  4.3ms
307.00 ms OK
RUN      T4 treesitter perf can redraw the end of a long line with wrapping:
Total min, 25%, median, 75%, max:
        1.4ms,  1.5ms,  1.6ms,  1.9ms,  4.7ms
334.00 ms OK
RUN      T5 treesitter perf can redraw the beginning of a long line without wrapping:
Total min, 25%, median, 75%, max:
        0.1ms,  0.1ms,  0.1ms,  0.2ms,  2.9ms
197.00 ms OK
RUN      T6 treesitter perf can redraw the middle of a long line without wrapping:
Total min, 25%, median, 75%, max:
        0.5ms,  0.6ms,  0.6ms,  0.7ms,  4.5ms
232.00 ms OK
RUN      T7 treesitter perf can redraw the end of a long line without wrapping:
Total min, 25%, median, 75%, max:
        0.9ms,  0.9ms,  1.0ms,  1.0ms,  3.7ms
266.00 ms OK
```