@vanaigr 2024-12-09
Updated the API.

`iter_captures()` now has an optional `opts` table with begin and end columns, and the iterator also accepts `end_col`.
`iter_captures()` remains backwards-compatible, but that means default values (from node bounds) are only used for lines and not for columns.

Docs for `on_range()` are not yet written since I'm not sure what the contract should be. Currently, the characters are considered included if their _first byte_ in inside the range, and begin/end positions can be on any byte. This means that if begin position is in the middle of a character, it is not included; and if end position is in the middle of a character, it IS included.

This seems convoluted, but that's the easiest way, and treesitter query cursor also interprets the bounds this way AFAIK.

There's also a theoretical problem with this approach if a treesitter node can start in the middle of a sequence of composing characters.

Should I make it so that range bounds always point to a first byte?