@vanaigr 2024-12-09
Added tests and news for `iter_captures` columns.