@vanaigr 2025-02-26
Rebased.

For line concealing, optimization is disabled and lines are checked fully.