@lewis6991 2025-08-03
> Could we deprecate on_line or does it have a strong purpose still?

Tbh I don't think it does.

Would be good to get a review from either @luukvbaal , @zeertzjq or @bfredl